+++ src/wave.h
@@ -1,7 +1,9 @@
+/* #include <stdint.h> */
+
 void wave_init();
-float wave_sine(int off);
-float wave_halfsine(int off);
-float wave_camelsine(int off);
-float wave_quartersine(int off);
-float wave_pulse(int off);
-float wave_ramp(int off);
+float wave_sine(uint16_t off);
+float wave_halfsine(uint16_t off);
+float wave_camelsine(uint16_t off);
+float wave_quartersine(uint16_t off);
+float wave_pulse(uint16_t off);
+float wave_ramp(uint16_t off);
